api/requirements.txt
pytest==8.3.3
pytest-html==4.1.1
requests==2.32.3
python-dotenv==1.0.1
